added a doc aand made components for everything
