+++ components/UpdateUser.tsx
@@ -1,0 +1,148 @@
+import { useState } from "react";
+
+function UpdateUserTester() {
+  const [userId, setUserId] = useState("");
+  const [name, setName] = useState("");
+  const [role, setRole] = useState("user");
+  const [subdomain, setSubdomain] = useState("");
+  const [companyDetails, setCompanyDetails] = useState("");
+  const [data, setData] = useState(null);
+  const [loading, setLoading] = useState(false);
+  const [error, setError] = useState("");
+
+  const handleUpdateUser = async (e: React.FormEvent) => {
+    e.preventDefault();
+    setLoading(true);
+    setData(null);
+    setError("");
+
+    try {
+      const response = await fetch("/api/update-user", {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify({
+          user_id: userId,
+          update_value: {
+            name,
+            role,
+            subdomain,
+            company_details: companyDetails,
+          },
+        }),
+      });
+
+      const result = await response.json();
+      if (!response.ok) {
+        throw new Error(result.message || "An error occurred");
+      }
+      setData(result);
+    } catch (err) {
+      if (err instanceof Error) {
+        setError(err.message);
+      }
+
+      console.error(err);
+    } finally {
+      setLoading(false);
+    }
+  };
+
+  return (
+    <div className="flex flex-col items-center p-6 bg-gray-100 min-h-screen">
+      <h1 className="text-2xl font-bold mb-4">
+        Test: /api/update-user Endpoint
+      </h1>
+      <form
+        onSubmit={handleUpdateUser}
+        className="w-full max-w-lg bg-white shadow-md rounded px-8 pt-6 pb-8 mb-4"
+      >
+        <div className="mb-4">
+          <label className="block text-gray-700 text-sm font-bold mb-2">
+            User ID:
+          </label>
+          <input
+            type="text"
+            value={userId}
+            onChange={(e) => setUserId(e.target.value)}
+            required
+            className="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
+          />
+        </div>
+        <div className="mb-4">
+          <label className="block text-gray-700 text-sm font-bold mb-2">
+            Name:
+          </label>
+          <input
+            type="text"
+            value={name}
+            onChange={(e) => setName(e.target.value)}
+            className="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
+          />
+        </div>
+        <div className="mb-4">
+          <label className="block text-gray-700 text-sm font-bold mb-2">
+            Role:
+          </label>
+          <select
+            value={role}
+            onChange={(e) => setRole(e.target.value)}
+            className="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
+          >
+            <option value="user">User</option>
+            <option value="admin">Admin</option>
+            <option value="compliance_manager">Compliance Manager</option>
+          </select>
+        </div>
+        <div className="mb-4">
+          <label className="block text-gray-700 text-sm font-bold mb-2">
+            Subdomain:
+          </label>
+          <input
+            type="text"
+            value={subdomain}
+            onChange={(e) => setSubdomain(e.target.value)}
+            className="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
+          />
+        </div>
+        <div className="mb-4">
+          <label className="block text-gray-700 text-sm font-bold mb-2">
+            Company Details:
+          </label>
+          <input
+            type="text"
+            value={companyDetails}
+            onChange={(e) => setCompanyDetails(e.target.value)}
+            className="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
+          />
+        </div>
+        <button
+          type="submit"
+          disabled={loading}
+          className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded focus:outline-none focus:shadow-outline w-full"
+        >
+          {loading ? "Updating..." : "Update User"}
+        </button>
+      </form>
+
+      {loading && <p className="text-gray-600">Loading...</p>}
+      {error && (
+        <div className="bg-red-100 border border-red-400 text-red-700 px-4 py-3 rounded relative mb-4 w-full max-w-[90vw]">
+          <h2 className="font-bold">Error</h2>
+          <pre>{error}</pre>
+        </div>
+      )}
+      {data && (
+        <div className="bg-green-100 border border-green-400 text-green-700 px-4 py-3 rounded relative mb-4 w-full max-w-[90vw]">
+          <h2 className="font-bold">Result from API</h2>
+          <pre className="whitespace-pre-wrap break-all">
+            {JSON.stringify(data, null, 2)}
+          </pre>
+        </div>
+      )}
+    </div>
+  );
+}
+
+export default UpdateUserTester;
